* fix library name formatting broken during rebase * point to Eigen fork that builds on OS X - pending Eigen PR merge
