+++ tensorflow/core/platform/load_library.h
@@ -25,6 +25,10 @@ namespace internal {
 Status LoadLibrary(const char* library_filename, void** handle);
 Status GetSymbolFromLibrary(void* handle, const char* symbol_name,
                             void** symbol);
+// Return the filename of a dynamically linked library formatted according to
+// platform naming conventions
+string FormatLibraryFileName(const string& name, const string& version);
+
 }  // namespace internal
 
 }  // namespace tensorflow
