+++ tensorflow/core/platform/posix/port.cc
@@ -83,24 +83,6 @@ void AdjustFilenameForLogging(string* filename) {
   // Nothing to do
 }
 
-string FormatLibraryFileName(const string& name, const string& version) {
-    string file_name;
-#if defined(__APPLE__)
-    if (version.size() == 0) {
-      file_name = "lib" + name + ".dylib";
-    } else {
-      file_name = "lib" + name + "." + version + ".dylib";
-    }
-#else
-    if (version.size() == 0) {
-      file_name = "lib" + name + ".so";
-    } else {
-      file_name = "lib" + name + ".so" + "." + version;
-    }
-#endif
-    return file_name;
-}
-
 bool Snappy_Compress(const char* input, size_t length, string* output) {
 #ifdef SNAPPY
   output->resize(snappy::MaxCompressedLength(length));
